fixes based on cr
