+++ kerem-net/src/Components/Post/Post.tsx
@@ -5,6 +5,7 @@ import List from '../List/List'
 import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
 import Reply from '../Reply/Reply';
 import ReplayModel from "../../Models/reply"
+
 interface PostProps{
     text:string
     authorName:string
