+++ kerem-net/src/Models/post.ts
@@ -1,0 +1,10 @@
+import reply from "./reply"
+
+interface PostData { text: string
+    authorName: string
+    likes: number
+    uploadTime: string
+    replies:reply[]
+}
+
+export default PostData;
